Translate payment received alert
