+++ lib/l10n/app_es.arb
@@ -1465,6 +1465,8 @@
       }
     }
   },
+  "successful_payment_print": "Imprimir",
+  "successful_payment_received": "¡Pago recibido!",
   "locale": "es",
   "app_name": "Breez"
 }
