+++ lib/l10n/app_fi.arb
@@ -1465,6 +1465,8 @@
       }
     }
   },
+  "successful_payment_print": "Print",
+  "successful_payment_received": "Payment received!",
   "locale": "fi",
   "app_name": "Breez"
 }
