+++ lib/l10n/app_pt.arb
@@ -1465,6 +1465,8 @@
       }
     }
   },
+  "successful_payment_print": "Imprimir",
+  "successful_payment_received": "Pagamento recebido!",
   "locale": "pt",
   "app_name": "Breez"
 }
